Refactoring PaymentValidator.java and PaymentValidatorImpl.java to CardValidator.java and CardValidatorImpl.java
