--- src/main/java/dev/andrylat/bankingutilities/card/CardValidator.java
@@ -1,9 +1,0 @@
-package dev.andrylat.bankingutilities.card;
-
-import java.util.List;
-
-public interface CardValidator {
-
-    List<String> validate(String customerInput);
-
-}
