--- src/main/java/dev/andrylat/bankingutilities/card/PaymentValidator.java
@@ -1,8 +1,0 @@
-package dev.andrylat.bankingutilities.card;
-
-
-public interface PaymentValidator {
-
-    String customerCardType();
-
-}
