+++ src/main/java/dev/andrylat/bankingutilities/card/validators/CardValidator.java
@@ -1,0 +1,13 @@
+package dev.andrylat.bankingutilities.card.validators;
+
+import java.util.List;
+
+public interface CardValidator {
+
+    List<String> validate(String customerInput);
+
+    String customerCardType();
+
+     void setPaymentCompanyIdentifier(int paymentCompanyIdentifier);
+
+}
