+++ src/test/java/dev/andrylat/bankingutilities/card/CardValidatorImplTest.java
@@ -1,5 +1,7 @@
 package dev.andrylat.bankingutilities.card;
 
+import dev.andrylat.bankingutilities.card.validators.CardValidator;
+import dev.andrylat.bankingutilities.card.validators.CardValidatorImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
